requirements.txt
torch
torchvision
torchaudio
matplotlib==3.9.0
librosa
gitpython